requirements.txt
# Core dependencies
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.4
python-jose==3.4.0
passlib==1.7.4
python-multipart==0.0.20
sqlalchemy==2.0.41
aiosmtplib==4.0.1
redis==6.1.0
httpx==0.28.1
email-validator==2.1.1

# Security
pyjwt==2.10.1
ecdsa==0.19.1
rsa==4.9.1
cryptography>=41.0.0

# Testing
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-cov==4.1.0
pytest-mock==3.14.0

# Development tools
black==24.3.0
isort==5.13.2
mypy==1.9.0
ruff==0.7.0

# Utilities
typing-extensions==4.13.2
anyio==4.9.0
python-dotenv==1.0.1

# Computer Vision
numpy>=1.26.0
opencv-python>=4.11.0
scikit-learn>=1.6.0
scipy>=1.15.0

# Database drivers
psycopg2-binary>=2.9.9  # PostgreSQL driver
aiomysql>=0.2.0  # MySQL async driver (if needed)

# Monitoring
prometheus-client>=0.19.0 